@@ -2,38 +2,37 @@ source 'https://rubygems.org'
 
 
 # Bundle edge Rails instead: gem 'rails', github: 'rails/rails'
-gem 'rails', '4.1.7'
-gem 'bootstrap-sass', '2.3.2.0'
+gem 'rails', '4.2.0'
+gem 'bootstrap-sass', '3.3.1.0'
 gem 'devise'
 gem 'omniauth'
 gem 'omniauth-facebook'
 # Use sqlite3 as the database for Active Record
 group :development, :test do 
- gem 'sqlite3', '1.3.8'
- gem 'rspec-rails', '2.13.1'
+ gem 'sqlite3', '1.3.10'
+ gem 'rspec-rails', '3.1.0'
 end
 
 group :test do
- gem 'selenium-webdriver', '2.35.1'
- gem 'capybara', '2.1.0'	
+ gem 'selenium-webdriver', '2.44.0'
+ gem 'capybara', '2.4.4'	
 end
 # Use SCSS for stylesheets
-gem 'sass-rails', '4.0.3'
+gem 'sass-rails', '5.0.1'
 # Use Uglifier as compressor for JavaScript assets
-gem 'uglifier', '1.3.0'
+gem 'uglifier', '2.7.0'
 # Use CoffeeScript for .js.coffee assets and views
-gem 'coffee-rails', '4.0.0'
+gem 'coffee-rails', '4.1.0'
 # See https://github.com/sstephenson/execjs#readme for more supported runtimes
 # gem 'therubyracer',  platforms: :ruby
 # Use jquery as the JavaScript library
-gem 'jquery-rails'
+gem 'jquery-rails', '4.0.3'
 # Turbolinks makes following links in your web application faster. Read more: https://github.com/rails/turbolinks
-gem 'turbolinks'
 # Build JSON APIs with ease. Read more: https://github.com/rails/jbuilder
-gem 'jbuilder', '2.0'
+gem 'jbuilder', '2.2.6'
 # bundle exec rake doc:rails generates the API under doc/api.
 group :doc do
- gem 'sdoc', '0.3.20', require: false
+ gem 'sdoc', '0.4.1', require: false
 end
 # Spring speeds up development by keeping your application running in the background. Read more: https://github.com/rails/spring
 group :production do
