@@ -1,6 +1,6 @@
 class SessionsController < ApplicationController
   def destroy
   	session[:user_id] = nil
-  	redirect_to root_url, :notice => "Log out"
+  	redirect_to root_path, :notice: "Log out"
   end
 end
