@@ -1,7 +1,7 @@
 Rails.application.routes.draw do
   root 'static_pages#home'
   devise_for :users, controllers: { omniauth_callbacks: 'users/omniauth_callbacks' }
-  match '/logout' => 'sessions#destroy', :as => :logout, via: 'get'
+  get '/logout' => 'sessions#destroy', as: :logout
   # The priority is based upon order of creation: first created -> highest priority.
   # See how all your routes lay out with "rake routes".
 
